factor out non-exported functions in sync & async actions to a utils file
