+++ lib/redux/actions/async.js
@@ -15,12 +15,11 @@
  * == BSD2 LICENSE ==
  */
 
-/* global chrome, __TEST__ */
+/* global chrome */
 
 import _ from 'lodash';
 import async from 'async';
 import semver from 'semver';
-import stacktrace from 'stack-trace';
 
 import sundial from 'sundial';
 
@@ -122,17 +121,13 @@ export function doAppInit(config, servicesToInit) {
         const { uploadTargetUser } = getState();
         if (uploadTargetUser !== null) {
           dispatch(syncActions.setBlipViewDataUrl(
-            api.makeBlipUrl(viewDataPathForUser(uploadTargetUser))
+            api.makeBlipUrl(actionUtils.viewDataPathForUser(uploadTargetUser))
           ));
         }
         dispatch(retrieveTargetsFromStorage());
       });
     });
   };
-}
-
-function viewDataPathForUser(uploadTargetUser) {
-  return `/patients/${uploadTargetUser}/data`;
 }
 
 export function doLogin(creds, opts) {
@@ -161,7 +156,7 @@ export function doLogin(creds, opts) {
       const { uploadTargetUser } = getState();
       if (uploadTargetUser !== null) {
         dispatch(syncActions.setBlipViewDataUrl(
-          api.makeBlipUrl(viewDataPathForUser(uploadTargetUser))
+          api.makeBlipUrl(actionUtils.viewDataPathForUser(uploadTargetUser))
         ));
       }
       dispatch(retrieveTargetsFromStorage());
@@ -183,50 +178,6 @@ export function doLogout() {
         dispatch(syncActions.setPage(pages.LOGIN, actionSources.USER));
       }
     });
-  };
-}
-
-function makeProgressFn(dispatch, step, percentage) {
-  return (step, percentage) => {
-    dispatch(syncActions.uploadProgress(step, percentage));
-  };
-}
-
-function makeUploadCb(dispatch, getState, errCode, utc) {
-  return (err, recs) => {
-    const { devices, uploadsByUser, uploadTargetDevice, uploadTargetUser, version } = getState();
-    const targetDevice = devices[uploadTargetDevice];
-    if (err) {
-      // the drivers sometimes just pass a string arg as err, instead of an actual error :/
-      if (typeof err === 'string') {
-        err = new Error(err);
-      }
-      const serverErr = 'Origin is not allowed by Access-Control-Allow-Origin';
-      let displayErr = new Error(err.message === serverErr ?
-        errorText.E_SERVER_ERR : errorText[errCode]);
-      let uploadErrProps = {
-        details: err.message,
-        utc: actionUtils.getUtc(utc),
-        name: err.name || 'Uncaught or API POST error',
-        step: err.step || null,
-        code: errCode,
-        version: version
-      };
-
-      if (!__TEST__) {
-        uploadErrProps.stringifiedStack = _.pluck(
-          _.filter(
-            stacktrace.parse(err),
-            (cs) => { return cs.functionName !== null; }
-          ),
-          'functionName'
-        ).join(', ');
-      }
-      return dispatch(syncActions.uploadFailure(displayErr, uploadErrProps, targetDevice));
-    }
-    const currentUpload = _.get(uploadsByUser, [uploadTargetUser, targetDevice.key], {});
-    dispatch(syncActions.uploadSuccess(uploadTargetUser, targetDevice, currentUpload, recs, utc));
-
   };
 }
 
@@ -271,10 +222,10 @@ export function doCareLinkUpload(deviceKey, creds, utc) {
       const opts = {
         targetId: uploadTargetUser,
         timezone: targetTimezones[uploadTargetUser],
-        progress: makeProgressFn(dispatch),
+        progress: actionUtils.makeProgressFn(dispatch),
         version: version
       };
-      carelink.upload(data, opts, makeUploadCb(dispatch, getState, 'E_CARELINK_UPLOAD', utc));
+      carelink.upload(data, opts, actionUtils.makeUploadCb(dispatch, getState, 'E_CARELINK_UPLOAD', utc));
     });
   };
 }
@@ -288,7 +239,7 @@ export function doDeviceUpload(driverId, utc) {
     const opts = {
       targetId: uploadTargetUser,
       timezone: targetTimezones[uploadTargetUser],
-      progress: makeProgressFn(dispatch),
+      progress: actionUtils.makeProgressFn(dispatch),
       version: version
     };
     const { uploadsByUser } = getState();
@@ -338,7 +289,7 @@ export function doDeviceUpload(driverId, utc) {
         return dispatch(syncActions.uploadFailure(displayErr, disconnectedErrProps, targetDevice));
       }
 
-      device.upload(driverId, opts, makeUploadCb(dispatch, getState, 'E_DEVICE_UPLOAD', utc));
+      device.upload(driverId, opts, actionUtils.makeUploadCb(dispatch, getState, 'E_DEVICE_UPLOAD', utc));
     });
   };
 }
@@ -467,12 +418,6 @@ export function doVersionCheck() {
  * COMPLEX ACTION CREATORS
  */
 
-function getDeviceTargetsByUser(targetsByUser) {
-  return _.mapValues(targetsByUser, (targets) => {
-    return _.pluck(targets, 'key');
-  });
-}
-
 export function putTargetsInStorage() {
   return (dispatch, getState) => {
     const { targetDevices, targetTimezones, uploadTargetUser } = getState();
@@ -504,7 +449,7 @@ export function putTargetsInStorage() {
       dispatch(syncActions.setPage(pages.MAIN));
     }
 
-    const uploadsByUser = getDeviceTargetsByUser(usersWithTargets);
+    const uploadsByUser = actionUtils.getDeviceTargetsByUser(usersWithTargets);
 
     dispatch(syncActions.setUploads(uploadsByUser));
   };
@@ -520,7 +465,7 @@ export function retrieveTargetsFromStorage() {
       return dispatch(syncActions.setPage(pages.SETTINGS));
     }
     else {
-      const uploadsByUser = getDeviceTargetsByUser(targets);
+      const uploadsByUser = actionUtils.getDeviceTargetsByUser(targets);
 
       dispatch(syncActions.setUploads(uploadsByUser));
     }
@@ -560,7 +505,7 @@ export function setUploadTargetUserAndMaybeRedirect(targetId) {
     dispatch(syncActions.setUploadTargetUser(targetId));
     const { api } = services;
     dispatch(syncActions.setBlipViewDataUrl(
-      api.makeBlipUrl(viewDataPathForUser(targetId))
+      api.makeBlipUrl(actionUtils.viewDataPathForUser(targetId))
     ));
     const targetedDevices = _.get(targetDevices, targetId, []);
     const targetedTimezone = _.get(targetTimezones, targetId, null);
