+++ lib/redux/actions/sync.js
@@ -300,17 +300,6 @@ export function fetchCareLinkFailure(message) {
  * relating to async action creator doUpload
  */
 
-function getUploadTrackingId(device) {
-  const source = device.source;
-  if (source.type === 'device' || source.type === 'block') {
-    return source.driverId;
-  }
-  if (source.type === 'carelink') {
-    return 'CareLink';
-  }
-  return null;
-}
-
 export function uploadAborted() {
   return {
     type: actionTypes.UPLOAD_ABORTED,
@@ -328,7 +317,7 @@ export function uploadRequest(userId, device, utc) {
     meta: {
       source: actionSources[actionTypes.UPLOAD_REQUEST],
       metric: {
-        eventName: `${metrics.UPLOAD_REQUEST} ${getUploadTrackingId(device)}`,
+        eventName: `${metrics.UPLOAD_REQUEST} ${actionUtils.getUploadTrackingId(device)}`,
         properties: {
           type: _.get(device, 'source.type', undefined),
           source: _.get(device, 'source.driverId', undefined)
@@ -354,7 +343,7 @@ export function uploadSuccess(userId, device, upload, data, utc) {
     meta: {
       source: actionSources[actionTypes.UPLOAD_SUCCESS],
       metric: {
-        eventName: `${metrics.UPLOAD_SUCCESS} ${getUploadTrackingId(device)}`,
+        eventName: `${metrics.UPLOAD_SUCCESS} ${actionUtils.getUploadTrackingId(device)}`,
         properties: {
           type: _.get(device, 'source.type', undefined),
           source: _.get(device, 'source.driverId', undefined),
@@ -376,7 +365,7 @@ export function uploadFailure(err, errProps, device) {
     meta: {
       source: actionSources[actionTypes.UPLOAD_FAILURE],
       metric: {
-        eventName: `${metrics.UPLOAD_FAILURE} ${getUploadTrackingId(device)}`,
+        eventName: `${metrics.UPLOAD_FAILURE} ${actionUtils.getUploadTrackingId(device)}`,
         properties: {
           type: _.get(device, 'source.type', undefined),
           source: _.get(device, 'source.driverId', undefined),
